+++ R/connect_hub.R
@@ -25,12 +25,14 @@
 #' `admin.json` and is ignored by default.
 #' If supplied, it will override hub configuration setting. Multiple formats can
 #' be supplied to `connect_hub` but only a single file format can be supplied to `connect_mod_out`.
-#' @param skip_checks Logical. If `FALSE` (default), check file_format parameter against the
-#' hub's model output files. Also excludes invalid model output files when opening hub datasets.
-#' Setting to `TRUE`` will improve performance but will result in an error if the model output
-#' directory includes invalid files. Cannot be `TRUE` when there are multiple file formats in
-#' the hub's model output directory or when the hub's model output directory contains files that
-#' are not model output data (for example, a README).
+#' @param skip_checks Logical. If `FALSE` (default), check `file_format` parameter
+#' against the hub's model output files. Also excludes invalid model output files
+#'  when opening hub datasets.
+#' Setting to `TRUE` will improve performance but will result in an error if
+#' the model output directory includes invalid files. **Cannot be `TRUE`** when
+#' there are **multiple file formats** in the hub's model output directory or when
+#' the hub's model output directory contains **files that are not model output
+#' data** (for example, a README).
 #' @inheritParams create_hub_schema
 #'
 #' @return